+++ src/Call/CallVerifier.php
@@ -26,6 +26,7 @@ use Eloquent\Phony\Event\Exception\UndefinedEventException;
 use Eloquent\Phony\Matcher\Matcher;
 use Eloquent\Phony\Matcher\MatcherFactory;
 use Eloquent\Phony\Matcher\MatcherVerifier;
+use Eloquent\Phony\Mock\Handle\InstanceHandle;
 use Eloquent\Phony\Verification\AbstractCardinalityVerifier;
 use Eloquent\Phony\Verification\GeneratorVerifier;
 use Eloquent\Phony\Verification\GeneratorVerifierFactory;
@@ -828,6 +829,10 @@ class CallVerifier extends AbstractCardinalityVerifier implements Call
                 return $this->assertionRecorder->createSuccess($matchingEvents);
             }
         } elseif (is_object($type)) {
+            if ($type instanceof InstanceHandle) {
+                $type = $type->get();
+            }
+
             if ($type instanceof Throwable || $type instanceof Exception) {
                 $isTypeSupported = true;
                 $type = $this->matcherFactory->equalTo($type, true);
@@ -873,6 +878,10 @@ class CallVerifier extends AbstractCardinalityVerifier implements Call
     public function threw($type = null)
     {
         $cardinality = $this->cardinality;
+
+        if ($type instanceof InstanceHandle) {
+            $type = $type->get();
+        }
 
         if ($type instanceof Throwable || $type instanceof Exception) {
             $type = $this->matcherFactory->equalTo($type, true);